requirements_ui.txt
absl-py==2.1.0
affine==2.3.0
aiohttp==3.7.4.post0
albumentations==1.3.1
aom
async-timeout==3.0.1
atk-1.0==2.38.0
attrs==22.1.0
blas==1.0
blinker==1.6.3
blosc==1.21.5
boost-cpp==1.78.0
branca==0.7.2
brotli==1.1.0
brotli-bin==1.1.0
brotli-python==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2024.3.11
cachetools==5.3.3
cairo==1.16.0
certifi==2024.2.2
cffi==1.15.1
cfitsio==4.1.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
cryptography==38.0.2
cuda-version==11.8
cudatoolkit==11.8.0
cudnn==8.9.7.29
curl==7.86.0
cycler==0.11.0
cython==0.29.32
cytoolz==0.12.0
dask-core==2022.2.0
decorator==4.4.2
easydict==1.9
exceptiongroup==1.2.0
execnet==2.0.2
expat==2.6.2
ffmpeg==4.4.2
fftw==3.3.10
fiona==1.8.21
flit-core==3.6.0
folium==0.16.0
font-ttf-dejavu-sans-mono==2.37
font-ttf-inconsolata==3.000
font-ttf-source-code-pro==2.038
font-ttf-ubuntu==0.83
fontconfig==2.14.2
fonts-conda-ecosystem==1
fonts-conda-forge==1
fonttools==4.38.0
freeglut==3.2.2
freetype==2.12.1
freexl==1.0.6
fribidi==1.0.10
fsspec==2023.1.0
future==0.18.2
gdal==3.5.0
gdk-pixbuf==2.42.8
geopandas==0.10.2
geopandas-base==0.10.2
geos==3.10.3
geotiff==1.7.1
gettext==0.22.5
gettext-tools==0.22.5
giflib==5.2.1
glib==2.78.4
glib-tools==2.78.4
google-auth==2.17.3
google-auth-oauthlib==0.4.6
graphite2==1.3.13
graphviz==7.0.5
grpc-cpp==1.48.1
grpcio==1.48.1
gst-plugins-base==1.20.2
gstreamer==1.20.3
gts==0.7.6
harfbuzz==5.3.0
hdf4==4.2.15
hdf5==1.12.1
icu==70.1
idna==3.7
imagecodecs-lite==2019.12.3
imageio==2.13.1
imgaug==0.4.0
importlib-metadata==4.11.3
importlib_metadata==4.11.3
iniconfig==2.0.0
jasper==2.0.33
jinja2==3.1.3
joblib==1.3.2
jpeg==9e
json-c==0.16
kealib==1.4.15
kiwisolver==1.3.2
krb5==1.19.3
lame==3.100
lcms2==2.12
lerc==3.0
libabseil==20220623.0
libasprintf==0.22.5
libasprintf-devel==0.22.5
libblas==3.9.0
libbrotlicommon==1.1.0
libbrotlidec==1.1.0
libbrotlienc==1.1.0
libcblas==3.9.0
libclang==14.0.6
libclang13==14.0.6
libcurl==7.86.0
libdb==6.2.32
libdeflate==1.10
libevent==2.1.10
libffi==3.4.4
libflac==1.3.4
libgd==2.3.3
libgdal==3.5.0
libgettextpo==0.22.5
libgettextpo-devel==0.22.5
libglib==2.78.4
libgomp==13.2.0
libhwloc==2.8.0
libiconv==1.16
libkml==1.3.0
liblapack==3.9.0
liblapacke==3.9.0
libllvm13==13.0.1
libllvm14==14.0.6
libnetcdf==4.8.1
libogg==1.3.4
libopenblas==0.3.21
libopencv==4.6.0
libopus==1.3.1
libpng==1.6.39
libpq==14.5
libprotobuf==3.20.1
librsvg==2.54.4
librttopo==1.1.0
libsndfile==1.0.31
libspatialindex==1.9.3
libspatialite==5.0.1
libsqlite==3.45.3
libssh2==1.10.0
libtiff==4.4.0
libvorbis==1.3.7
libwebp==1.2.4
libwebp-base==1.2.4
libxcb==1.13
libxml2==2.10.2
libzip==1.8.0
libzlib==1.2.13
llvm-openmp==18.1.3
locket==1.0.0
lz4-c==1.9.4
magma==2.5.4
mapclassify==2.5.0
markdown==3.6
markupsafe==2.1.1
matplotlib==3.5.3
matplotlib-base==3.5.3
mkl==2022.2.1
mock==5.1.0
multidict==6.0.2
munch==2.5.0
munkres==1.1.4
mysql-common==8.0.32
mysql-libs==8.0.32
networkx==2.5.1
ninja==1.12.0
numpy==1.21.5
numpy-base==1.21.5
oauthlib==3.2.2
olefile==0.47
opencv==4.6.0
openh264==2.3.1
openjpeg==2.5.0
openssl==1.1.1w
packaging==23.2
pandas==1.3.5
pango==1.50.12
partd==1.4.1
path==16.5.0
path.py==12.5.0
pcre==8.45
pcre2==10.42
pillow==9.2.0
pip==22.3.1
pixman==0.40.0
pluggy==1.0.0
poppler==22.04.0
poppler-data==0.4.11
postgresql==14.5
proj==9.0.0
protobuf==3.20.1
pthread-stubs==0.4
py-opencv==4.6.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyjwt==2.8.0
pyopenssl==23.2.0
pyparsing==3.0.9
pyproj==3.2.1
pyqt==5.15.4
pyqt5-sip==12.9.0
pyshp==2.3.1
pysocks==1.7.1
pytest==7.4.4
pytest-shutil==1.7.0
python==3.7.16
python-dateutil==2.9.0
python_abi==3.7
pytorch==1.12.1
pytz==2024.1
pyu2f==0.1.5
pywavelets==1.3.0
pyyaml==6.0
qt-main==5.15.4
qudida==0.0.4
rasterio==1.2.10
re2==2022.06.01
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==59.8.0
shapely==1.8.2
sip==6.5.1
six==1.16.0
sleef==3.5.1
snappy==1.2.0
snuggs==1.4.7
sqlite==3.45.3
sridentify==0.8.0
svt-av1==1.3.0
tbb==2021.7.0
tensorboard==2.11.2
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2020.6.3
tiledb==2.9.5
tk==8.6.12
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.2
tqdm==4.66.2
typing-extensions==4.7.1
typing_extensions==4.7.1
tzdata==2024a
unicodedata2==14.0.0
uriparser==0.9.7
urllib3==2.2.1
werkzeug==2.2.3
wheel==0.38.4
x264==1!164.3095
x265==3.5
xerces-c==3.2.4
xorg-fixesproto==5.0
xorg-inputproto==2.3.2
xorg-kbproto==1.0.7
xorg-libice==1.1.1
xorg-libsm==1.2.4
xorg-libx11==1.8.4
xorg-libxau==1.0.11
xorg-libxdmcp==1.1.3
xorg-libxext==1.3.4
xorg-libxfixes==5.0.3
xorg-libxi==1.7.10
xorg-libxrender==0.9.10
xorg-renderproto==0.11.1
xorg-xextproto==7.3.0
xorg-xproto==7.0.31
xyzservices==2023.5.0
xz==5.4.6
yaml==0.2.5
yarl==1.7.2
zipp==3.11.0
zlib==1.2.13
zstd==1.5.5
libtasn1==4.19.0
libpciaccess==0.18
libgfortran5==11.2.0
libidn2==2.3.7
libudev1==253
libvpx==1.11.0
libnsl==2.0.1
readline==8.2
torchvision==0.13.0
pulseaudio==14.0
gnutls==3.7.9
gmp==6.3.0
nccl==2.21.5.1
libdrm==2.4.120
jack==1.9.18
libgfortran-ng==11.2.0
gtk2==2.24.33
libdap4==3.20.6
libgcc-ng==13.2.0
libxkbcommon==1.0.3
libnghttp2==1.52.0
libstdcxx-ng==13.2.0
tzcode==2024a
nss==3.98
alsa-lib==1.2.3.2
libva==2.18.0
libcap==2.64
p11-kit==0.24.1
ld_impl_linux-64==2.38
libuuid==2.38.1
libedit==3.1.20230828
libglu==9.0.0
attr==2.5.1
libev==4.33
nettle==3.9.1
dbus==1.13.6
libunistring==0.9.10
keyutils==1.6.1
nspr==4.35
libtool==2.4.7
libcups==2.3.3
ncurses==6.4